chore: moved some things around
